Re-use MaturesOnCell for open shorts
