+++ apps/hyperdrive-trading/src/ui/portfolio/OpenLongsTable/OpenLongsTable.tsx
@@ -20,7 +20,7 @@ import { formatBalance } from "src/ui/base/formatting/formatBalance";
 import { useReadHyperdrive } from "src/ui/hyperdrive/hooks/useReadHyperdrive";
 import { CloseLongModalButton } from "src/ui/hyperdrive/longs/CloseLongModalButton/CloseLongModalButton";
 import { usePreviewCloseLong } from "src/ui/hyperdrive/longs/hooks/usePreviewCloseLong";
-import { useCurrentBlock } from "src/ui/network/useCurrentBlock";
+import { MaturesOnCell } from "src/ui/portfolio/MaturesOnCell/MaturesOnCell";
 import { parseUnits } from "viem";
 import { useAccount } from "wagmi";
 interface OpenLongsTableProps {
@@ -39,9 +39,7 @@ function getColumns({ hyperdrive }: { hyperdrive: Hyperdrive }) {
       id: "maturationDate",
       header: `Matures on`,
       cell: ({ row }) => {
-        return <MaturesOnCell row={row} />;
-        // const maturity = new Date(Number(row.original.maturity * 1000n));
-        // return <span>{maturity.toLocaleDateString()}</span>;
+        return <MaturesOnCell maturity={row.original.maturity} />;
       },
     }),
     columnHelper.accessor("bondAmount", {
@@ -101,31 +99,6 @@ function getColumns({ hyperdrive }: { hyperdrive: Hyperdrive }) {
   ];
 }
 
-function MaturesOnCell({ row }: { row: Row<Long> }) {
-  const { currentBlock } = useCurrentBlock();
-  const isTermComplete =
-    row.original.maturity < (currentBlock?.timestamp || 0n);
-  const maturityDateMS = row.original.maturity * 1000n;
-  const termEndDate = new Date(Number(maturityDateMS));
-  const daysLeft = convertMillisecondsToDays(
-    Number(maturityDateMS - (currentBlock?.timestamp || 0n) * 1000n),
-  );
-
-  return (
-    <div className="flex flex-col gap-1">
-      <span className="ml-2">{termEndDate.toLocaleDateString()}</span>
-      {isTermComplete ? (
-        <div className={"daisy-badge daisy-badge-md inline-flex"}>
-          <span>{"Term complete"}</span>
-        </div>
-      ) : (
-        <div className={"daisy-badge daisy-badge-md inline-flex"}>
-          <span>{daysLeft} days left</span>
-        </div>
-      )}
-    </div>
-  );
-}
 function FixedRateCell({
   row,
   hyperdrive,
